FormulaTypes can now calculate themselves through Table refference that is given to them during initialization phase.

+++ Project/include/IntegerType.h
@@ -4,7 +4,7 @@
 class IntegerType : public CellType
 {
 private:
-	int m_value;
+	long long m_value;
 
 public:
 	/**
